feat(web): public viewer editor access

+++ apps/web/features/sidebar/ui/navigation-item-menu.tsx
@@ -1,3 +1,4 @@
+import type { DocumentProps } from "@lir/lib/schema";
 import {
   Icons,
   Button,
@@ -7,11 +8,13 @@ import {
   DropdownMenuTrigger,
 } from "@lir/ui";
 
+import { createId } from "@paralleldrive/cuid2";
 import { useMutation } from "@tanstack/react-query";
 
 import { usePathname, useRouter } from "next/navigation";
 
 import { documentApi, documentModel } from "~/entities/document";
+import { sessionModel } from "~/entities/session";
 import { generateDocumentURL } from "~/shared";
 
 type Props = {
@@ -24,45 +27,61 @@ export const NavigationItemMenu = ({ itemId, open, setOpen }: Props) => {
   const router = useRouter();
   const pathname = usePathname();
 
+  const isAuth = sessionModel.useAuth();
   const currentDocument = documentModel.useCurrentDocument();
   const documents = documentModel.useDocuments();
   const documentIndex = documentModel
     .useDocuments()
     .findIndex((document) => document.id === itemId);
 
+  const cleanupDocument = (deletedId: string) => {
+    documentModel.unsetDocument({ id: deletedId });
+
+    if (deletedId === currentDocument?.id) {
+      documentModel.setCurrentDocument(null);
+    }
+  };
+
   const { mutateAsync: deleteDocument } = useMutation({
     mutationKey: documentApi.documentKeys.mutation.updateDocument(),
     mutationFn: documentApi.deleteDocument,
-    onSuccess: (deletedDocument) => {
-      // queryClient.invalidateQueries({
-      //   queryKey: documentApi.documentKeys.query.getUserDocuments(),
-      // });
-
-      documentModel.unsetDocument({
-        id: deletedDocument.id,
-      });
-
-      if (deletedDocument.id === currentDocument?.id) {
-        documentModel.setCurrentDocument(null);
-      }
-    },
+    onSuccess: (deletedDocument) => cleanupDocument(deletedDocument.id),
   });
 
   const handleDeleteDocument = async () => {
-    await deleteDocument({ id: itemId });
-
-    if (documents.length === 1) {
-      if (pathname.startsWith("/d/")) {
-        router.replace("/d");
-      }
-
-      return;
+    if (isAuth) {
+      await deleteDocument({ id: itemId });
+    } else {
+      cleanupDocument(itemId);
     }
 
-    const nextDocument =
-      documents.length === documentIndex + 1
-        ? documents[documentIndex - 1]
-        : documents[documentIndex + 1];
+    let nextDocument: DocumentProps;
+
+    if (!isAuth) {
+      nextDocument = {
+        id: createId(),
+        title: "Untitled Document",
+        userId: "",
+        content: {},
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      };
+
+      documentModel.setDocuments([nextDocument]);
+    } else {
+      if (documents.length === 1) {
+        if (pathname.startsWith("/d/")) {
+          router.replace("/d");
+        }
+
+        return;
+      }
+
+      nextDocument =
+        documents.length === documentIndex + 1
+          ? documents[documentIndex - 1]
+          : documents[documentIndex + 1];
+    }
 
     router.replace(generateDocumentURL(nextDocument.title, nextDocument.id));
   };
